requirements.txt
##### dependency ###########
Django==2.0.7
django-crispy-forms==1.7.2
django-markdown-deux==1.0.5
django-pagedown==1.0.6
markdown2==2.3.7
Pillow==5.4.1
psycopg2==2.7.4
pytz==2018.9
#############################
############# Django-notifications ##########
django-notifications-hq==1.5.0
## automatic installed for Django-notification
django-model-utils==3.1.2
jsonfield==2.0.2
#############################################
